refactor: move XLSX file reading logic to a separate utility module and update related imports
